Add files via upload
HumanEval is the place where this thing performs exceedingly well, yet on there the algorithm is extremely redundant, the MCTS is essentially not there, and the entire thing just reduce to trial-and-error type iteration. Then this algorithm basically runs 140 test and pick out the best one from there.
Significant improvement room is possible here ... and the efficiency of the algorithm itself is in doubt. But, action + observation, evaluation score might be a good idea, we basically want more and more context, as our semantic feedback, and the MCTS part is simply not cracked well in the paper, too.